speed up initialisation

+++ core/functions.sh
@@ -162,8 +162,11 @@ common_init(){
     force_permissive=$(iniparser /etc/debian.conf "default" "force_permissive")
     if [[ ${force_permissive} != "false" ]] ; then
         setenforce 0 &>/dev/null || true
+        sed -i "s/^SELINUX=*/SELINUX=disabled/" /etc/sysconfig/selinux &>/dev/null || true
     fi
-    chmod +x /usr/lib/sulin/dsl/* 
+    if [[ -f ${DESTDIR}/run/debian ]] ; then
+        return
+    fi
     cp -prf /usr/lib/sulin/dsl/debrun ${DESTDIR}/bin/debrun
     cp -prf /usr/lib/sulin/dsl/hostctl ${DESTDIR}/bin/hostctl
     cp -prf /usr/lib/sulin/dsl/debxdg ${DESTDIR}/bin/debxdg
